fix: ReJest beforeEach blocks
