+++ apps/common-app/src/apps/reanimated/examples/RuntimeTests/tests/runtimes/runOnRuntimeAsync.test.tsx
@@ -20,12 +20,8 @@ describe('runOnRuntimeAsync', () => {
     notify(FAIL_NOTIFICATION);
   };
 
-  test('setup beforeEach', () => {
-    // TODO: there's a bug in ReJest and beforeEach has to be registered
-    // inside a test case.
-    beforeEach(() => {
-      reason = '';
-    });
+  beforeEach(() => {
+    reason = '';
   });
 
   test('schedules on RN Runtime to a Worker Runtime', async () => {
