+++ apps/common-app/src/apps/reanimated/examples/RuntimeTests/tests/runtimes/runOnUISync.test.tsx
@@ -25,13 +25,9 @@ describe('runOnUISync', () => {
     notify(FAIL_NOTIFICATION);
   };
 
-  test('setup beforeEach', () => {
-    // TODO: there's a bug in ReJest and beforeEach has to be registered
-    // inside a test case.
-    beforeEach(() => {
-      value = 0;
-      reason = '';
-    });
+  beforeEach(() => {
+    value = 0;
+    reason = '';
   });
 
   test('schedules on RN Runtime to UI Runtime', () => {
