+++ apps/common-app/src/apps/reanimated/examples/RuntimeTests/tests/runtimes/scheduleOnRN.test.tsx
@@ -12,12 +12,8 @@ describe('scheduleOnRN', () => {
     notify(PASS_NOTIFICATION);
   };
 
-  test('setup beforeEach', () => {
-    // TODO: there's a bug in ReJest and beforeEach has to be registered
-    // inside a test case.
-    beforeEach(() => {
-      value = 0;
-    });
+  beforeEach(() => {
+    value = 0;
   });
 
   test('schedules on RN Runtime to RN Runtime', async () => {
