+++ apps/common-app/src/apps/reanimated/examples/RuntimeTests/tests/runtimes/scheduleOnUI.test.tsx
@@ -19,13 +19,9 @@ describe('scheduleOnUI', () => {
     notify(FAIL_NOTIFICATION);
   };
 
-  test('setup beforeEach', () => {
-    // TODO: there's a bug in ReJest and beforeEach has to be registered
-    // inside a test case.
-    beforeEach(() => {
-      value = 0;
-      reason = '';
-    });
+  beforeEach(() => {
+    value = 0;
+    reason = '';
   });
 
   test('schedules on RN Runtime to UI Runtime', async () => {
